policy: Change Identity and Nexthdr types
Key Identity and Nexthdr fields are now defined as
identity.NumericIdentity and u8proto.U8proto, respectively. This improves
type safety and almost eliminates the need for type casts regarding
these.

+++ pkg/envoy/xds_server.go
@@ -1624,7 +1624,7 @@ func getDirectionNetworkPolicy(ep endpoint.EndpointUpdater, l4Policy policy.L4Po
 
 		port := l4.Port
 		if port == 0 && l4.PortName != "" {
-			port = ep.GetNamedPort(l4.Ingress, l4.PortName, uint8(l4.U8Proto))
+			port = ep.GetNamedPort(l4.Ingress, l4.PortName, l4.U8Proto)
 			if port == 0 {
 				return true // Skip if a named port can not be resolved (yet)
 			}
